requirements.txt
pandas
snowflake-connector-python
matplotlib
